Bugfix: TVL set1 primitives used wrong base type.

+++ include/vector/simd/avx512/primitives/create_avx512.h
@@ -47,7 +47,7 @@ namespace vectorlib {
       template< typename U = T, typename std::enable_if< std::is_integral< U >::value, int >::type = 0 >
       MSV_CXX_ATTRIBUTE_FORCE_INLINE
       static typename avx512< v512< U > >::vector_t
-      set1( int a0) {
+      set1( uint64_t a0) {
          trace( "[VECTOR] - set1 sse register." );
          return _mm512_set1_epi64(a0);
       }
